ULAPI fixes
- Remove sources in liblinuxcnculapi.so from liblinuxcnchal.so and
  link executables against liblinuxcnculapi.so
- Make ULAPI even more like other flavors
  - Shares the `rtapi_app_main`
  - Has its own `flavor_descriptor`
  - `ulapi_main.c` and `rtapi_main.c` merged & folded into
    `rtapi_common.c`
- Fix (remove) ULAPI autoloading
  - Still gets initialized at first `rtapi_init()` call
- Clean out MODULE_EXT build macro, obsolete since kthreads dumped
