+++ src/Makefile
@@ -238,12 +238,12 @@ TOOBJSDEPS = $(call TOOBJS,$(1),$(2)) $(call TODEPS, $(1),$(2))
 # $$(eval $(call c_comp_build_rules,hal/components/encoder.o,hal/components/othersrc.o))
 # obj-m += encoder.o
 # encoder-objs := hal/components/encoder.o hal/components/othersrc.o
-# $(RTLIBDIR)/encoder$(MODULE_EXT): \
+# $(RTLIBDIR)/encoder.so: \
 #     hal/components/encoder.o hal/components/othersrc.o
 define c_comp_build_rules
 obj-m += $$(patsubst $$(dir $(1))%.o,%.o,$(1))
 $$(patsubst $$(dir $(1))%.o,%-objs,$(1)) := $(1) $(2)
-$$(patsubst $$(dir $(1))%.o,$(RTLIBDIR)/%$(MODULE_EXT),$(1)): \
+$$(patsubst $$(dir $(1))%.o,$(RTLIBDIR)/%.so,$(1)): \
     $$(addprefix $(OBJDIR)/,$(1) $(2))
 
 endef
@@ -596,7 +596,6 @@ HEADERS := \
     rtapi/rtapi_compat.h \
     rtapi/rtapi_hexdump.h \
     rtapi/rtapi_int.h \
-    rtapi/rtapi_kdetect.h \
     rtapi/rtapi_limits.h \
     rtapi/rtapi_math.h \
     rtapi/rtapi_math64.h \
@@ -609,14 +608,17 @@ HEADERS := \
     rtapi/rtapi_errno.h \
     rtapi/rtapi_string.h \
     rtapi/rtapi_pci.h \
-    rtapi/rtapi_proc.h \
     rtapi/rtapi_heap.h \
     rtapi/rtapi_heap_private.h \
     rtapi/ring.h \
     rtapi/triple-buffer.h \
     rtapi/multiframe.h \
     rtapi/rtapi_mbarrier.h \
-    rtapi/shmdrv/shmdrv.h
+    rtapi/shmdrv/shmdrv.h \
+    rtapi/flavor/rtapi_flavor.h \
+    rtapi/flavor/xenomai.h \
+    rtapi/flavor/rt-preempt.h \
+    rtapi/flavor/ulapi.h
 
 
 
@@ -715,7 +717,7 @@ clean: depclean modclean docclean
 	-rm -rf ../rtlib ../libexec
 	rm -f ../etc/linuxcnc/rtapi.ini
 	-rm -f $(COPY_CONFIGS)
-	-rm -f $(RTLIBDIR)/*.$(MODULE_EXT)
+	-rm -f $(RTLIBDIR)/*.so
 	-rm -f hal/components/conv_*.comp
 	-rm -f hal/i_components/conv_*.icomp
 	-rm -f hal/components/*_bin.h   # generated by pasm for bb platform
@@ -1082,6 +1084,7 @@ EXTRA_CFLAGS = $(RTFLAGS) \
 	-I$(INCLUDE_PREFIX)/libnml/buffer \
 	-I$(INCLUDE_PREFIX)/libnml/posemath \
 	-I$(INCLUDE_PREFIX)/rtapi \
+	-I$(INCLUDE_PREFIX)/rtapi/flavor \
 	-I$(INCLUDE_PREFIX)/hal/lib \
 	-I$(INCLUDE_PREFIX)/emc/nml_intf \
 	-I$(INCLUDE_PREFIX)/emc/kinematics \
@@ -1192,7 +1195,7 @@ RTDEPS := $(sort $(patsubst $(OBJDIR)/%.o,$(DEPDIR)/%.d, $(RTOBJS)))
 
 modules:
 ifeq ($(RUN_IN_PLACE)+$(BUILD_DRIVERS),yes+yes)
-	test -f ../libexec/rtapi_app -a \
+	@test -f ../libexec/rtapi_app -a \
 	    \( 0`stat -c %u ../libexec/rtapi_app 2>/dev/null` \
 			-ne 0 -o ! -u ../libexec/rtapi_app \) \
 		&& need_setuid=1; \
@@ -1308,9 +1311,13 @@ objects/var-%: Makefile $(wildcard $(SUBMAKEFILES)) Makefile.inc
 	@echo $($*) > $@.tmp
 	@sh move-if-change $@.tmp $@
 
+# Link shared libs
 ../lib/%.so: ../lib/%.so.0
 	$(ECHO) Symlinking $(notdir $<) to $(notdir $@)
 	$(Q)ln -sf $(notdir $<) $@
+# - Be sure the setuid message comes last
+ALL_SHLIBS=$(filter ../%.so.0,$(TARGETS)) $(filter ../%.so,$(TARGETS))
+modules: $(ALL_SHLIBS)
 
 cscope:
 	cscope -Rb
