+++ src/Makefile.inc.in
@@ -78,8 +78,6 @@ RUN_IN_PLACE = @RUN_IN_PLACE@
 RTFLAGS := -I. -DRTAPI -D_GNU_SOURCE -D_FORTIFY_SOURCE=0 $(THREADS_RTFLAGS)
 ULFLAGS = -D_GNU_SOURCE -DLOCALE_DIR=\"$(localedir)\" -DPACKAGE=\"$(package)\"
 
-MODULE_EXT = .so
-
 CC = @CC@
 CC_FOR_BUILD = @CC_FOR_BUILD@
 OBJCOPY = @OBJCOPY@
