+++ src/emc/iotask/Submakefile
@@ -2,11 +2,23 @@ IOSRCS := emc/iotask/ioControl.cc emc/rs274ngc/tool_parse.cc
 IOV2SRCS := emc/iotask/ioControl_v2.cc emc/rs274ngc/tool_parse.cc
 USERSRCS += $(IOSRCS) $(IOV2SRCS)
 
-../bin/io: $(call TOOBJS, $(IOSRCS)) ../lib/liblinuxcnc.a ../lib/libnml.so.0 ../lib/liblinuxcnchal.so.0 ../lib/liblinuxcncini.so.0
+../bin/io: \
+		$(call TOOBJS, $(IOSRCS)) \
+		../lib/liblinuxcnc.a \
+		../lib/libnml.so.0 \
+		../lib/liblinuxcnchal.so.0 \
+		../lib/liblinuxcnculapi.so.0 \
+		../lib/liblinuxcncini.so.0
 	$(ECHO) Linking $(notdir $@)
 	@$(CXX) $(LDFLAGS) -o $@ $^ 
 
-../bin/iov2: $(call TOOBJS, $(IOV2SRCS)) ../lib/liblinuxcnc.a ../lib/libnml.so.0 ../lib/liblinuxcnchal.so.0 ../lib/liblinuxcncini.so.0
+../bin/iov2: \
+		$(call TOOBJS, $(IOV2SRCS)) \
+		../lib/liblinuxcnc.a \
+		../lib/libnml.so.0 \
+		../lib/liblinuxcnchal.so.0 \
+		../lib/liblinuxcnculapi.so.0 \
+		../lib/liblinuxcncini.so.0
 	$(ECHO) Linking $(notdir $@)
 	@$(CXX) $(LDFLAGS) -o $@ $^
