+++ src/hal/classicladder/Submakefile
@@ -44,7 +44,10 @@ $(call TOOBJS, $(CLASSICLADDERSRCS)): Makefile.inc
 
 USERSRCS += $(CLASSICLADDERSRCS)
 
-../bin/classicladder: $(call TOOBJS, $(CLASSICLADDERSRCS)) ../lib/liblinuxcnchal.so.0
+../bin/classicladder: \
+		$(call TOOBJS, $(CLASSICLADDERSRCS)) \
+		../lib/liblinuxcnchal.so.0 \
+		../lib/liblinuxcnculapi.so.0
 	$(ECHO) Linking $(notdir $@)
 	@$(CC) $(LDFLAGS) $(CFLAGS) $(ULFLAGS) -o $@ $^ $(GTK_LIBS) -lpthread
