+++ src/rtapi/rtapi_app.h
@@ -28,16 +28,10 @@
 #if defined(USERMODE_PCI)
 #include "userpci/module.h"
 #endif
-/*  For kernel modules (hm2_pci, hostmot2) to compile in usermode without lots
-    of changes, the EXPORT_SYMBOL lines, below, need to be defined *ONLY* if
-    there is actually an rtapi_app_* function in the code.  This is handeled
-    in the kernel module case by the simple define that renames the module
-    init/cleanup functions.
-    For the usermode case, turn the first instance of rtapi_app_* into a
-    function declaration, then export the symbol, then re-create the function
-    definition.  This way the symbol is only exported if the function exists
-    in the code, and we don't have to have the ability to 'rewind' the C
-    preprocessor.
+/*  Turn the first instance of rtapi_app_* into a function declaration, then
+    export the symbol, then re-create the function definition.  This way the
+    symbol is only exported if the function exists in the code, and we don't
+    have to have the ability to 'rewind' the C preprocessor.
  */
 #define rtapi_app_main(a)           \
     rtapi_app_main(a);              \
